Consolidate WebSocketController and ActiveGameController
